+++ blam.lua
@@ -1,9 +1,9 @@
 ------------------------------------------------------------------------------
 -- Blam! library for Chimera/SAPP Lua scripting
 -- Sledmine, JerryBrick
--- Improves memory handle and provides standard functions for scripting
+-- Easier memory handle and provides standard functions for scripting
 ------------------------------------------------------------------------------
-local blam = {_VERSION = "1.4.0"}
+local blam = {_VERSION = "1.5.0-beta"}
 
 ------------------------------------------------------------------------------
 -- Useful functions for internal usage
@@ -43,11 +43,11 @@ end
 -- Engine address list
 local addressList = {
     tagDataHeader = 0x40440000,
-    cameraType = 0x00647498, -- from Giraffe
+    cameraType = 0x00647498, --from giraffe
     gamePaused = 0x004ACA79,
     gameOnMenus = 0x00622058,
-    joystickInput = 0x64D998, -- from aLTis
-    firstPerson = 0x40000EB8, -- from aLTis
+    joystickInput = 0x64D998, --from aLTis
+    firstPerson = 0x40000EB8, --from aLTis
     objectTable = 0x400506B4,
     deviceGroupsTable = 0x00816110
 }
@@ -168,7 +168,7 @@ local cameraTypes = {
     deadCamera = 5 -- 23776
 }
 
--- Netgame flags type 
+-- Netgame flags type
 local netgameFlagTypes = {
     ctfFlag = 0,
     ctfVehicle = 1,
@@ -251,29 +251,31 @@ local dPadValues = {
     up = 765
 }
 
-------------------------------------------------------------------------------
--- SAPP API bindings
-------------------------------------------------------------------------------
--- All the functions at the top of the module are for EmmyLua autocompletion purposes!
--- They do not have a real implementation and are not supossed to be imported
+-- EmmyLua autocompletion for some functions!
+-- Functions below do not have a real implementation and are not supossed to be imported
 if (variableThatObviouslyDoesNotExist) then
-
-    --- Attempt to spawn an object given tag id and coordinates or tag type and class plus coordinates
-    ---@param tagId number Optional tag id of the object to spawn
+    ---Attempt to spawn an object given tag class and path plus coordinates
+    ---@overload fun(tagId: string, x: number, y: number, z: number):number
     ---@param tagType string Type of the tag to spawn
     ---@param tagPath string Path of object to spawn
     ---@param x number
     ---@param y number
     ---@param z number
+    ---@return number | nil objectId
     function spawn_object(tagType, tagPath, x, y, z)
     end
 
-    --- Get object address from a specific player given playerIndex
+    ---Get object address from a specific player given playerIndex
     ---@param playerIndex number
-    ---@return number Player object memory address
+    ---@return number objectAddress
     function get_dynamic_player(playerIndex)
     end
+
 end
+
+------------------------------------------------------------------------------
+-- SAPP API bindings
+------------------------------------------------------------------------------
 if (api_version) then
     -- Provide global server type variable on SAPP
     server_type = "sapp"
@@ -293,10 +295,10 @@ if (api_version) then
         return array
     end
 
-    --- Function wrapper for file writing from Chimera to SAPP
+    ---Write content to a text file given file path
     ---@param path string Path to the file to write
     ---@param content string Content to write into the file
-    ---@return boolean | nil, string True if successful otherwise nil, error
+    ---@return boolean | nil, string result True if successful otherwise nil, error
     function write_file(path, content)
         local file, error = io.open(path, "w")
         if (not file) then
@@ -312,9 +314,9 @@ if (api_version) then
         end
     end
 
-    --- Function wrapper for file reading from Chimera to SAPP
+    ---Read the contents from a file given file path
     ---@param path string Path to the file to read
-    ---@return string | nil, string Content of the file otherwise nil, error
+    ---@return string | nil, string content string if successful otherwise nil, error
     function read_file(path)
         local file, error = io.open(path, "r")
         if (not file) then
@@ -328,17 +330,21 @@ if (api_version) then
         return content
     end
 
-    -- TODO PENDING FUNCTION!!
-    function directory_exists(dir)
+    ---Verify if a directory exists given directory path
+    ---@param path string
+    ---@return boolean
+    function directory_exists(path)
+        error("Directory verifications are not supported on SAPP.. yet!")
         return true
     end
 
-    --- Function wrapper for directory listing from Chimera to SAPP
-    ---@param dir string
-    function list_directory(dir)
+    ---List the contents from a directory given directory path
+    ---@param path string
+    ---@return integer | table
+    function list_directory(path)
         -- TODO This needs a way to separate folders from files
-        if (dir) then
-            local command = "dir " .. dir .. " /B"
+        if (path) then
+            local command = "dir " .. path .. " /B"
             local pipe = io.popen(command, "r")
             local output = pipe:read("*a")
             if (output) then
@@ -354,7 +360,7 @@ if (api_version) then
         return nil
     end
 
-    --- Return the memory address of a tag given tagId or tagClass and tagPath
+    ---Return the memory address of a tag given tagId or tagClass and tagPath
     ---@param tagIdOrTagType string | number
     ---@param tagPath string
     ---@return number
@@ -366,13 +372,13 @@ if (api_version) then
         end
     end
 
-    --- Execute a game command or script block
+    ---Execute a game command or script block
     ---@param command string
     function execute_script(command)
         return execute_command(command)
     end
 
-    --- Return the address of the object memory given object id
+    ---Return the address of the object memory given object id
     ---@param objectId number
     ---@return number
     function get_object(objectId)
@@ -391,40 +397,42 @@ if (api_version) then
         destroy_object(objectId)
     end
 
-    --- Print text into console
+    ---Print text into console
     ---@param message string
     ---@param red number
     ---@param green number
     ---@param blue number
     function console_out(message, red, green, blue)
-        -- TODO Add color printing to this function
+        -- TODO Add color printing to this function on SAPP
         cprint(message)
     end
 
-    --- Get if the game console is opened \
-    --- Always returns true on SAPP.
+    ---Get if the game console is opened, always returns true on SAPP
     ---@return boolean
     function console_is_open()
         return true
     end
 
-    --- Get the value of a Halo scripting global\
-    ---An error will occur if the global is not found.
+    ---Get the value of a Halo scripting global.\
+    ---An error will be triggered if the global is not found
     ---@param name string Name of the global variable to get from hsc
     ---@return boolean | number
     function get_global(name)
-        error("SAPP can't retrieve global variables as Chimera does.. yet!")
+        error("SAPP can not retrieve global variables as Chimera does.. yet!")
     end
 
-    --- Print messages to the player HUD\
-    ---Server messages will be printed if executed from SAPP.
+    ---Print message to player HUD.\
+    ---Messages will be printed to console if SAPP uses this function
     ---@param message string
     function hud_message(message)
-        for playerIndex = 1, 16 do
-            if (player_present(playerIndex)) then
-                rprint(playerIndex, message)
-            end
-        end
+        cprint(message)
+    end
+
+    ---Set the callback for an event game from the game events available on Chimera
+    ---@param event '"command"' | '"frame"' | '"preframe"' | '"map_load"' | '"precamera"' | '"rcon message"' | '"tick"' | '"pretick"' | '"unload"'
+    ---@param callback string global function name to call when the event is triggered
+    function set_callback(event, callback)
+        error("Chimera events can not be used on SAPP, use register_callback instead.")
     end
 
     print("Compatibility with Chimera Lua API has been loaded!")
@@ -714,7 +722,7 @@ end
 
 local function readList(address, propertyData)
     local operation = typesOperations[propertyData.elementsType]
-    local elementCount = read_byte(address - 0x4)
+    local elementCount = read_word(address - 0x4)
     local addressList = read_dword(address) + 0xC
     if (propertyData.noOffset) then
         addressList = read_dword(address)
@@ -1367,6 +1375,14 @@ local weaponHudInterfaceStructure = {
 ---@field type number
 ---@field teamIndex number
 
+---@class cutsceneFlag
+---@field name string
+---@field x number
+---@field y number
+---@field z number
+---@field vX number
+---@field vY number
+
 ---@class scenario
 ---@field sceneryPaletteCount number Number of sceneries in the scenery palette
 ---@field sceneryPaletteList table Tag ID list of scenerys in the scenery palette
@@ -1380,6 +1396,8 @@ local weaponHudInterfaceStructure = {
 ---@field netgameFlagsList table List of netgame equipments
 ---@field objectNamesCount number Count of the object names in the scenario
 ---@field objectNames string[] List of all the object names in the scenario
+---@field cutsceneFlagsCount number Count of all the cutscene flags in the scenario
+---@field cutsceneFlags cutsceneFlag[] List of all the cutscene flags in the scenario
 
 -- Scenario structure
 local scenarioStructure = {
@@ -1457,6 +1475,20 @@ local scenarioStructure = {
         elementsType = "string",
         jump = 36,
         noOffset = true
+    },
+    cutsceneFlagsCount = {type = "dword", offset = 0x4E4},
+    cutsceneFlags = {
+        type = "table",
+        offset = 0x4E8,
+        jump = 92,
+        rows = {
+            name = {type = "string", offset = 0x4},
+            x = {type = "float", offset = 0x24},
+            y = {type = "float", offset = 0x28},
+            z = {type = "float", offset = 0x2C},
+            vX = {type = "float", offset = 0x30},
+            vY = {type = "float", offset = 0x34}
+        }
     }
 }
 
@@ -1641,6 +1673,10 @@ local projectileStructure = extendStructure(objectStructure, {
 ---@field speed number Current speed of this player
 ---@field ping number Ping amount from server of this player in milliseconds
 ---@field kills number Kills quantity done by this player
+---@field assists number Assists count of this player
+---@field betraysAndSuicides number Betrays plus suicides count of this player
+---@field deaths number Deaths count of this player
+---@field suicides number Suicides count of this player
 
 local playerStructure = {
     id = {type = "word", offset = 0x0},
@@ -1652,7 +1688,11 @@ local playerStructure = {
     index = {type = "byte", offset = 0x67},
     speed = {type = "float", offset = 0x6C},
     ping = {type = "dword", offset = 0xDC},
-    kills = {type = "word", offset = 0x9C}
+    kills = {type = "word", offset = 0x9C},
+    assists = {type = "word", offset = 0XA4},
+    betraysAndSuicides = {type = "word", offset = 0xAC},
+    deaths = {type = "word", offset = 0xAE},
+    suicides = {type = "word", offset = 0XB0}
 }
 
 ---@class firstPersonInterface
@@ -1690,7 +1730,7 @@ local firstPersonStructure = {weaponObjectId = {type = "dword", offset = 0x10}}
 ---@field disableCollision number Disable collision of this biped tag
 local bipedTagStructure = {disableCollision = {type = "bit", offset = 0x2F4, bitLevel = 5}}
 
----@class  deviceMachine : blamObject
+---@class deviceMachine : blamObject
 ---@field powerGroupIndex number Power index from the device groups table
 ---@field power number Position amount of this device machine
 ---@field powerChange number Power change of this device machine
@@ -1738,6 +1778,30 @@ local hudGlobalsStructure = {
     textSpacing = {type = "float", offset = 0x90}
 }
 
+---@class widgetInstance
+---@field tagId number
+---@field name string
+---@field frameX number
+---@field frameY number
+---@field widgetType number
+---@field visible boolean
+---@field opacity number
+---@field previousWidget number
+---@field nextWidget number
+---@field parentWidget number
+---@field childWidget number
+---@field text string
+---@field cursorIndex number
+---@field focused boolean
+local widgetInstanceStructure = {
+    tagId = {type = "dword", offset = 0x0},
+    -- name = {type = "dword", offset = 0x4},
+    frameX = {type = "short", offset = 8},
+    frameY = {type = "short", offset = 12},
+    widgetType = {type = "word", offset = 16},
+    visible = {type = "bit", offset = 20, bitLevel = 0x0}
+}
+
 ------------------------------------------------------------------------------
 -- LuaBlam globals
 ------------------------------------------------------------------------------
@@ -1769,7 +1833,7 @@ blam.tagDataHeader = createObject(addressList.tagDataHeader, tagDataHeaderStruct
 blam.dumpObject = dumpObject
 blam.consoleOutput = consoleOutput
 
---- Get if a value equals a null value for game
+--- Get if a value equals a null value in game terms
 ---@return boolean
 function blam.isNull(value)
     if (value == 0xFF or value == 0xFFFF or value == 0xFFFFFFFF or value == nil) then
@@ -1791,10 +1855,10 @@ function blam.isGameDedicated()
 end
 
 function blam.isGameSAPP()
-    return server_type == "sapp"
+    return server_type == "sapp" or api_version
 end
 
---- Get the current game camera type
+---Get the current game camera type
 ---@return number
 function blam.getCameraType()
     local camera = read_word(addressList.cameraType)
@@ -1815,12 +1879,12 @@ function blam.getCameraType()
     return nil
 end
 
---- Get input from the joystick in the game
--- Based on aLTis controller method
--- TODO Check if it is better to return an entire table with all input values 
+--- Get input from joystick assigned in the game
 ---@param joystickOffset number Offset input from the joystick data, use blam.joystickInputs
 ---@return boolean | number Value of the joystick input
 function blam.getJoystickInput(joystickOffset)
+    -- Based on aLTis controller method
+    -- TODO Check if it is better to return an entire table with all input values 
     joystickOffset = joystickOffset or 0
     -- Nothing is pressed by default
     local inputValue = false
@@ -2172,13 +2236,60 @@ function blam.getDeviceGroup(index)
     -- Get object address
     if (index) then
         -- Get objects table
-        local table = createObject(read_dword(addressList.deviceGroupsTable), deviceGroupsTableStructure)
+        local table = createObject(read_dword(addressList.deviceGroupsTable),
+                                   deviceGroupsTableStructure)
         -- Calculate object ID (this may be invalid, be careful)
-        local itemOffset = table.elementSize * index 
+        local itemOffset = table.elementSize * index
         local item = read_float(table.firstElementAddress + itemOffset + 0x4)
         return item
     end
     return nil
+end
+
+---@class blamRequest
+---@field requestString string
+---@field timeout number
+---@field callback function<boolean, string>
+---@field sentAt number
+
+---@type table<number, blamRequest>
+local requestQueue = {}
+local requestId = -1
+local requestPathMaxLength = 60
+---Send a server request to current server trough rcon
+---@param method '"GET"' | '"SEND"'
+---@param url string Path or name of the resource we want to get
+---@param timeout number Time this request will wait for a response
+---@param callback function<boolean, string> Callback function to call when this response returns
+---@param retry boolean Retry this request if timeout reaches it's limit
+---@param params table<string, any> Optional parameters to send in the request, careful, this will create two requests, one for the resource and another one for the parameters
+---@return boolean success
+function blam.request(method, url, timeout, callback, retry, params)
+    if (not timeout) then
+        timeout = 120
+    end
+    if (server_type ~= "dedicated") then
+        console_out("Warning, requests only work while connected to a dedicated server.")
+    end
+    if (params) then
+        console_out("Warning, request params are not supported yet.")
+    end
+    if (url and url:len() <= requestPathMaxLength) then
+        if (method == "GET") then
+            requestId = requestId + 1
+            local rconRequest = ("rcon blam ?%s?%s"):format(requestId, url)
+            requestQueue[requestId] = {
+                requestString = rconRequest,
+                timeout = timeout,
+                callback = callback
+            }
+            console_out(rconRequest)
+            -- execute_script(request)
+            return true
+        end
+    end
+    error("Error, url can not contain more than " .. requestPathMaxLength .. " chars.")
+    return false
 end
 
 return blam